requirements.txt
crewai==0.41.1
exa_py==1.0.9
langchain==0.2.12
python-dotenv==1.0.1
langchain-core==0.2.29
crewai-tools==0.4.26
